ENH: created separate module for communication functions
